Add correct install steps for *nix
I tried the previous steps, but didn't work, so went to the AUR package to extract what it did (I'm not on AUR, I'm on Kubuntu 20.04), tested it, and it worked flawlessly. So, I thought updating the docs was a good idea :)  
I can expect the Win steps to be broken too, but can't test, so ¯\_(ツ)_/¯

Please make sure it's correct for other platform, if possible though.